Always on volume control on now pĺaying screen
Options for large and small album layout. Large album layouyt adds a
volume bar where the volume button used to be. Small album art displays
the volume control which were previously in a bottom sheet.

Move the button for the current playlist to the top bar, so it's always
the same place.