Fix HomeMenuActivity not updated when activity resumes
after archived item is restored or shortcut is added.

The previous implementation relied on getting a sticky HomeMenuEvent
which referred to a home menu object which contains the now updated home
menu entries. Withouyt the event actually being sent. Now we make sure
to send a livedata update when the home menu changes, which will be
delivered when the activity is in an active state, e.g. on resume.